% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_team_colour_palette.R
\name{extract_matchday_palette}
\alias{extract_matchday_palette}
\title{Extract NRL Match Day Colour Palettes}
\usage{
extract_matchday_palette(home_team = NULL, away_team = NULL)
}
\arguments{
\item{home_team}{An NRL team name.}

\item{away_team}{The number of colours to be extracted from the
colour palette.}
}
\value{
The NRL team colour palette.
}
\description{
Extract NRL Match Day Colour Palettes
}
\examples{
extract_matchday_palette("Melbourne Storm", 2)
extract_matchday_palette("Parramatta Eels", 3)
}
