% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_team_colour_palette.R
\name{extract_team_palette}
\alias{extract_team_palette}
\title{Extract NRL Team Colour Palette}
\usage{
extract_team_palette(team_name = NULL, n_colours = 1)
}
\arguments{
\item{team_name}{An NRL team name.}

\item{n_colours}{The number of colours to be extracted from the
colour palette.}
}
\value{
The NRL team colour palette.
}
\description{
Extract NRL Team Colour Palette
}
\examples{
extract_team_palette("Melbourne Storm", 2)
extract_team_palette("Parramatta Eels", 3)
}
