% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_team_colour_palette.R
\name{plot_team_palette}
\alias{plot_team_palette}
\title{Plot NRL Team Colour Palette}
\usage{
plot_team_palette(team_name)
}
\arguments{
\item{team_name}{An NRL team name.}
}
\value{
A visual of the NRL team colour palette.
}
\description{
Plot NRL Team Colour Palette
}
\examples{
\dontrun{
 plot_team_palette("Melbourne Storm")
}
\dontrun{
 plot_team_palette("Parramatta Eels")
}
}
